ej 5 y 6

+++ labos/python/prtactica7/ej3_1.py
@@ -1,4 +1,4 @@
-def alguno_es_0(n1: int, n2: int) -> bool:
+def alguno_es_0(n1: float, n2: float) -> bool:
     return n1 * n2 == 0
